+++ src/components/Button/Button.tsx
@@ -1,7 +1,7 @@
 import * as PropTypes from 'prop-types'
 import * as React from 'react'
 
-import { UIComponent, childrenExist, customPropTypes } from '../../lib'
+import { UIComponent, childrenExist, customPropTypes, createShorthandFactory } from '../../lib'
 import Icon from '../Icon'
 import { ButtonBehavior } from '../../lib/accessibility'
 import { Accessibility } from '../../lib/accessibility/interfaces'
@@ -12,6 +12,7 @@ import {
   ReactChildren,
   ComponentEventHandler,
 } from '../../../types/utils'
+import ButtonGroup from './ButtonGroup'
 
 export interface IButtonProps {
   as?: any
@@ -22,8 +23,10 @@ export interface IButtonProps {
   content?: React.ReactNode
   fluid?: boolean
   icon?: ItemShorthand
+  iconOnly?: boolean
   iconPosition?: 'before' | 'after'
   onClick?: ComponentEventHandler<IButtonProps>
+  text?: boolean
   type?: 'primary' | 'secondary'
   accessibility?: Accessibility
   styles?: IComponentPartStylesInput
@@ -42,6 +45,8 @@ export interface IButtonProps {
  *  - if button includes icon only, textual representation needs to be provided by using 'title', 'aria-label', or 'aria-labelledby' attributes
  */
 class Button extends UIComponent<Extendable<IButtonProps>, any> {
+  static create: Function
+
   public static displayName = 'Button'
 
   public static className = 'ui-button'
@@ -71,6 +76,9 @@ class Button extends UIComponent<Extendable<IButtonProps>, any> {
     /** Button can have an icon. */
     icon: customPropTypes.itemShorthand,
 
+    /** A button may indicate that it has only icon. */
+    iconOnly: PropTypes.bool,
+
     /** An icon button can format an Icon to appear before or after the button */
     iconPosition: PropTypes.oneOf(['before', 'after']),
 
@@ -80,6 +88,9 @@ class Button extends UIComponent<Extendable<IButtonProps>, any> {
      * @param {object} data - All props.
      */
     onClick: PropTypes.func,
+
+    /** A button can be formatted to show only text in order to indicate some less-pronounced actions. */
+    text: PropTypes.bool,
 
     /** A button can be formatted to show different levels of emphasis. */
     type: PropTypes.oneOf(['primary', 'secondary']),
@@ -104,9 +115,11 @@ class Button extends UIComponent<Extendable<IButtonProps>, any> {
     'disabled',
     'fluid',
     'icon',
+    'iconOnly',
     'iconPosition',
     'onClick',
     'styles',
+    'text',
     'type',
     'variables',
   ]
@@ -116,7 +129,15 @@ class Button extends UIComponent<Extendable<IButtonProps>, any> {
     accessibility: ButtonBehavior as Accessibility,
   }
 
-  public renderComponent({ ElementType, classes, accessibility, rest }): React.ReactNode {
+  static Group = ButtonGroup
+
+  public renderComponent({
+    ElementType,
+    classes,
+    accessibility,
+    variables,
+    rest,
+  }): React.ReactNode {
     const { children, content, disabled, iconPosition } = this.props
     const hasChildren = childrenExist(children)
 
@@ -130,19 +151,18 @@ class Button extends UIComponent<Extendable<IButtonProps>, any> {
       >
         {hasChildren && children}
         {!hasChildren && iconPosition !== 'after' && this.renderIcon()}
-        {!hasChildren && content}
+        {!hasChildren && content && <span className={classes.content}>{content}</span>}
         {!hasChildren && iconPosition === 'after' && this.renderIcon()}
       </ElementType>
     )
   }
 
   public renderIcon = () => {
-    const { disabled, icon, iconPosition, content, type } = this.props
+    const { icon, iconPosition, content } = this.props
 
     return Icon.create(icon, {
       defaultProps: {
         xSpacing: !content ? 'none' : iconPosition === 'after' ? 'before' : 'after',
-        variables: { color: type === 'primary' && !disabled ? 'white' : undefined },
       },
     })
   }
@@ -160,5 +180,7 @@ class Button extends UIComponent<Extendable<IButtonProps>, any> {
     }
   }
 }
+
+Button.create = createShorthandFactory(Button, content => ({ content }))
 
 export default Button