+++ src/components/Text/Text.tsx
@@ -3,10 +3,34 @@ import * as React from 'react'
 
 import { childrenExist, customPropTypes, UIComponent } from '../../lib'
 
+import { Extendable } from '../../../types/utils'
+import { ComponentVariablesInput, IComponentPartStylesInput } from '../../../types/theme'
+
+export interface ITextProps {
+  as?: any
+  atMention?: boolean
+  className?: string
+  content?: any
+  disabled?: boolean
+  error?: boolean
+  important?: boolean
+  size?: 'xs' | 'sm' | 'md' | 'lg' | 'xl' | '2x' | '3x' | '4x'
+  weight?: 'light' | 'semilight' | 'regular' | 'semibold' | 'bold'
+  success?: boolean
+  temporary?: boolean
+  timestamp?: boolean
+  truncated?: boolean
+  styles?: IComponentPartStylesInput
+  variables?: ComponentVariablesInput
+}
+
 /**
  * A component containing text
+ * @accessibility
+ * Text is how people read the content on your website.
+ * Ensure that a contrast ratio of at least 4.5:1 exists between text and the background behind the text.
  */
-class Text extends UIComponent<any, any> {
+class Text extends UIComponent<Extendable<ITextProps>, any> {
   static className = 'ui-text'
 
   static displayName = 'Text'
@@ -42,6 +66,9 @@ class Text extends UIComponent<any, any> {
     /** Set as success Text component */
     success: PropTypes.bool,
 
+    /** The text can signify a temporary state */
+    temporary: PropTypes.bool,
+
     /** Set as timestamp Text component */
     timestamp: PropTypes.bool,
 
@@ -70,6 +97,7 @@ class Text extends UIComponent<any, any> {
     'size',
     'styles',
     'success',
+    'temporary',
     'timestamp',
     'truncated',
     'variables',