+++ src/lib/accessibility/Behaviors/Menu/MenuItemBehavior.ts
@@ -1,5 +1,6 @@
 import { Accessibility } from '../../interfaces'
 import { IS_FOCUSABLE_ATTRIBUTE } from '../../FocusZone/FocusUtilities'
+import * as keyboardKey from 'keyboard-key'
 
 const MenuItemBehavior: Accessibility = (props: any) => ({
   attributes: {
@@ -8,8 +9,20 @@ const MenuItemBehavior: Accessibility = (props: any) => ({
     },
     anchor: {
       role: 'menuitem',
-      'aria-expanded': props['submenuOpened'],
+      tabIndex: '0',
+      'aria-label': props['aria-label'],
+      'aria-labelledby': props['aria-labelledby'],
       [IS_FOCUSABLE_ATTRIBUTE]: true,
+    },
+  },
+
+  handledProps: ['aria-label', 'aria-labelledby'],
+
+  keyActions: {
+    anchor: {
+      performClick: {
+        keyCombinations: [{ keyCode: keyboardKey.Enter }, { keyCode: keyboardKey.Spacebar }],
+      },
     },
   },
 })