+++ src/themes/teams/components/Text/textStyles.ts
@@ -2,15 +2,27 @@ import { Sizes, Weights } from '../../../../lib/enums'
 import { ICSSInJSStyle } from '../../../../../types/theme'
 import { truncateStyle } from '../../../../styles/customCSS'
 import { ITextVariables } from './textVariables'
+import { ITextProps } from '../../../../components/Text/Text'
 
 export interface TextStylesParams {
-  props: any
+  props: ITextProps
   variables: ITextVariables
 }
 
 export default {
   root: ({
-    props: { atMention, disabled, error, size, weight, important, success, timestamp, truncated },
+    props: {
+      atMention,
+      disabled,
+      error,
+      size,
+      weight,
+      important,
+      success,
+      timestamp,
+      truncated,
+      temporary,
+    },
     variables: v,
   }: TextStylesParams): ICSSInJSStyle => {
     return {
@@ -19,7 +31,13 @@ export default {
       ...(disabled && { color: v.disabledTextColor }),
       ...(error && { color: v.errorTextColor }),
       ...(success && { color: v.successTextColor }),
-      ...(timestamp && { color: v.timestampTextColor }),
+      ...(temporary && { fontStyle: 'italic' }),
+      ...(timestamp && {
+        color: v.timestampTextColor,
+        ':hover': {
+          color: v.timestampHoverTextColor,
+        },
+      }),
       ...(weight === Weights.Light && {
         fontWeight: v.textWeightLight,
       }),
@@ -35,7 +53,10 @@ export default {
       ...(weight === Weights.Bold && {
         fontWeight: v.textWeightBold,
       }),
-      ...(important && { fontWeight: v.importantWeight }),
+      ...(important && {
+        fontWeight: v.importantWeight,
+        color: v.importantTextColor,
+      }),
       ...(size === Sizes.ExtraSmall && {
         fontSize: v.textExtraSmallFontSize,
         lineHeight: v.textExtraSmallLineHeight,