+++ test/specs/components/Text/Text-test.tsx
@@ -1,0 +1,15 @@
+import * as React from 'react'
+import { shallow } from 'enzyme'
+
+import { isConformant } from 'test/specs/commonTests'
+import { mountWithProvider } from 'test/utils'
+
+import Text from 'src/components/Text'
+
+describe('Text', () => {
+  isConformant(Text)
+
+  test('renders children', () => {
+    expect(mountWithProvider(<Text>children</Text>).text()).toEqual('children')
+  })
+})